requirements.txt
asgiref==3.6.0
black @ git+https://github.com/psf/black@3feff21eca0eee4b9fe72d12b506ac273cb5bcd0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.4
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
